# openapi.yaml
openapi: 3.0.3
info:
  title: SoloBackend API
  description: |
    API for the SoloBackend digital music bundle store. Used by the web app and
    mobile clients (e.g. Soloband). Authentication uses Kinde JWT; send it as
    `Authorization: Bearer <access_token>` for protected endpoints.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://your-production-domain.com
    description: Production (replace with your base URL)
...
security:
  - BearerAuth: []
paths:
...
  /api/collection:
    get:
      tags: [Collection]
      summary: Get My Collection
      description: |
        Returns the merged collection: purchased + owned + saved bundles, deduplicated.
        Each item has `type` (purchased | owned | saved), `purchased`, and `unlocked`.
        Use `unlocked` with the bundle's internal lock metadata to control full vs partial content.
      responses:
        "200":
          description: Collection items
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/CollectionItem" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Collection]
      summary: Add bundle to collection
      description: Add a bundle to the user's collection (saved). Idempotent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bundle_id]
              properties:
                bundle_id:
                  type: string
                  format: uuid
      responses:
        "200":
          description: Added
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    properties:
                      added:
                        type: boolean
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/purchase/create-checkout-session:
    post:
      tags: [Purchase]
      summary: Create Stripe Checkout session
      description: Creates a Stripe Checkout session for the given bundle. Redirect the user to the returned `checkout_url`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bundle_id]
              properties:
                bundle_id:
                  type: string
                  format: uuid
                email:
                  type: string
                  format: email
                  description: Optional; used if JWT has no email claim
      responses:
        "200":
          description: Checkout URL
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: object
                    required: [checkout_url]
                    properties:
                      checkout_url:
                        type: string
                        format: uri
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Kinde access token (JWT)
  schemas:
    Bundle:
      type: object
      required: [id, title, description, price, category, created_at]
      properties:
        id:
          type: string
          format: uuid
        title: { type: string }
        description: { type: string }
        price: { type: number }
        category: { type: string }
        metadata_url:
          type: string
          format: uri
          nullable: true
        r2_key:
          type: string
          nullable: true
        created_by_user_id:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    CollectionItem:
      type: object
      required: [bundle, type, purchased, unlocked]
      properties:
        bundle: { $ref: "#/components/schemas/Bundle" }
        type:
          type: string
          enum: [purchased, owned, saved]
        purchased:
          type: boolean
          description: True if user has purchased this bundle
        unlocked:
          type: boolean
          description: True if user has full access (purchased or created); use with bundle lock metadata for full vs partial content
        purchased_at:
          type: string
          format: date-time
        added_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
  responses:
    Unauthorized:
      description: Missing or invalid Authorization header (Bearer token)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: unauthorized
            message: Invalid or expired token
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: not_found
            message: Bundle not found
    ValidationError:
      description: Invalid request (validation failed)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: validation
            message: title is required
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }